+++ crates/toasty-core/src/stmt/visit_mut.rs
@@ -1254,8 +1254,8 @@ where
             }
         }
         Returning::Changed => {}
+        Returning::Project(expr) => v.visit_expr_mut(expr),
         Returning::Expr(expr) => v.visit_expr_mut(expr),
-        Returning::Value(expr) => v.visit_expr_mut(expr),
     }
 }